YUK-11909 Web Schedules and Read Attribute need to support RF meters/review changes

+++ yukon-client/yukon-web/src/com/cannontech/web/amr/outageProcessing/OutageProcessingController.java
@@ -125,7 +125,7 @@ public class OutageProcessingController extends MultiActionController {
                 int total = (int)result.getOriginalDeviceCollectionCopy().getDeviceCount();
                 float percentSuccess = 100.0f;
                 if (total > 0) {
-                    percentSuccess = (successCount * 100) / total;
+                    percentSuccess = ((float) successCount * 100) / total;
                 }
                 resolvableTemplate.addData("percentSuccess", percentSuccess);
                 resolvableTemplate.addData("resultKey", result.getKey());
